+++ server/TWS Foundation/TWS Business.Quality/Q_Depots/Q_Drivers_Commons.cs
@@ -1,10 +1,11 @@
 ﻿using TWS_Business.Depots;
 using TWS_Business.Entities;
 using TWS_Business.Entities.Drivers;
+using TWS_Business.Quality.Q_Depots.Bases;
 
 namespace TWS_Business.Quality.Q_Depots;
 
-public class Q_Drivers_Commons : BQ_Business<Driver_Common, Drivers_CommonsDepot> {
+public class Q_Drivers_Commons : BQ_Common<Driver_Common, Driver, DriverExternal, Drivers_CommonsDepot> {
 
     protected override Driver_Common EntityFactory(string Entropy) {
 
@@ -40,17 +41,18 @@ public class Q_Drivers_Commons : BQ_Business<Driver_Common, Drivers_CommonsDepot
                   }
              );
 
-        DriverExternal external = Store(
-                new DriverExternal {
-                    Identification = identification,
-                }
-            );
-
-        return new Driver_Common {
+        //// Como hacer?
+        //// - Create hace un sanity y truena si no encuentra un external o internal.
+        ///
+        Driver_Common common = new Driver_Common {
             License = Entropy[..12],
             Situation = situation,
             Status = status,
-            External = external,
+            External = new DriverExternal {
+                Identification = identification,
+            }
         };
+      
+        return common;
     }
 }